refactor: consolidate the discovery walk on the fallow-core implementation
Delete crates/engine/src/discover_walk.rs, a near-verbatim fork of
crates/core/src/discover/walk.rs, and route the engine discovery path
through the core_backend adapter instead. The surviving core walk derives
its config-candidate basenames live from the plugin registry, so the
hardcoded BUILTIN_PLUGIN_CONFIG_CANDIDATE_BASENAMES list and its no-drift
parity gate are deleted with the fork.

Counts: 37 insertions, 1974 deletions (net -1937 lines).
